requirements.txt
amqp==5.0.6
anyjson==0.3.3
asgiref==3.3.4
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==5.0.5
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.7
defusedxml==0.7.1
Django==3.2.4
django-allauth==0.44.0
django-bootstrap4==3.0.1
django-celery-beat==2.2.0
django-cleanup==5.2.0
django-countries==7.2.1
django-environ==0.4.5
django-ranged-response==0.2.0
django-simple-captcha==0.5.14
django-timezone-field==4.1.2
idna==2.10
kombu==5.1.0
oauthlib==3.1.1
Pillow==8.2.0
prompt-toolkit==3.0.18
pycparser==2.20
PyJWT==2.1.0
python-crontab==2.5.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
